+++ __tests__/components/contribution/SubmitMessageInputForm.test.tsx
@@ -4,7 +4,7 @@ import { Provider } from "react-redux";
 // Note: test renderer must be required after react-native.
 import renderer from "react-test-renderer";
 import createMockStore from "redux-mock-store";
-import SubmitMessageInputForm from "../../../src/components/contribution/AddPostInputs";
+import SubmitMessageInputForm from "../../../src/components/contribution/SubmitMessageInputForm";
 import { IState } from "../../../src/state/state/IState";
 import { Pick2 } from "../../../src/utils/ts/Pick2";
 import TestLocalizationProvider from "../../helpers/TestLocalizationProvider";
