+++ src/utils/pickMessageContent.ts
